# Optimize provision speed a bit, focusing on 1-node GCP