--- a/sky/skylet/ray_patches/command_runner.py.patch
+++ b/sky/skylet/ray_patches/command_runner.py.patch
@@ -1,0 +1,4 @@
+330c330
+<                     "ControlPersist": "10s",
+---
+>                     "ControlPersist": "300s",