--- a/sky/utils/log_utils.py
+++ b/sky/utils/log_utils.py
@@ -1,4 +1,4 @@
-"""Sky logging utils."""
+"""Logging utils."""
 import enum
 from typing import List, Optional