… sky/templates/gcp-ray.yml.j2
… tests/test_smoke.py