# Before vs after
`sky launch --cloud gcp`

Before, master 054402e3d77 - median 3:02
```
  sky launch -c no-opt --cloud gcp -y ''  22.03s user 4.66s system 14% cpu 3:02.50 total
  sky launch -c no-opt1 --cloud gcp -y ''  19.38s user 3.81s system 12% cpu 3:06.32 total
  sky launch -c no-opt2 --cloud gcp -y ''  19.02s user 3.73s system 12% cpu 2:58.92 total
```
After - median 2:40 (12% speed up)
```
sky launch -c opt --cloud gcp -y ''  13.89s user 3.17s system 10% cpu 2:43.79 total
sky launch -c opt2 --cloud gcp -y ''  11.78s user 2.37s system 8% cpu 2:40.32 total
sky launch -c opt9 --cloud gcp -y ''  11.49s user 2.37s system 9% cpu 2:31.00 total
```

After, but **additionally comment out any credentials upload** (see below) - median 2:13 (27% speedup)
```
  sky launch -c test1 --cloud gcp -y ''  12.68s user 2.68s system 11% cpu 2:13.05 total
  sky launch -c test2 --cloud gcp -y ''  11.15s user 2.10s system 9% cpu 2:16.99 total
  sky launch -c test3 --cloud gcp -y ''  11.06s user 2.09s system 10% cpu 2:05.38 total
```

**Lesson**: **`file_mounts` in generated Ray yaml are very expensive**. 

Each mount = 1 SSH open/disconnect. We can consider 
- zipping them
- using an env var to turn off uploading

I decided to not do this as the target user may be GCP-only, so they may be better than what I have:
```yaml
file_mounts:
  ~/.sky/sky_ray.yml: /Users/zongheng/.sky/generated/opt2.yml
  ~/.sky/sky_wheels: /var/folders/8f/56gzvwkd3n3293xjlrztr6600000gp/T/08d9c1b43ad1c22d17325787d7dae62d
  ~/.aws/credentials: ~/.aws/credentials
  ~/.azure/azureProfile.json: ~/.azure/azureProfile.json
  ~/.azure/clouds.config: ~/.azure/clouds.config
  ~/.azure/config: ~/.azure/config
  ~/.azure/msal_token_cache.json: ~/.azure/msal_token_cache.json
  ~/.config/gcloud/credentials.db: ~/.config/gcloud/credentials.db
  ~/.config/gcloud/application_default_credentials.json: ~/.config/gcloud/application_default_credentials.json
  ~/.config/gcloud/access_tokens.db: ~/.config/gcloud/access_tokens.db
  ~/.config/gcloud/configurations: ~/.config/gcloud/configurations
  ~/.config/gcloud/legacy_credentials: ~/.config/gcloud/legacy_credentials

```

# Breakdown of current overheads

Note for future: **The best way to see these breakdowns** is to read the various timestamps in 
```
I 08-17 13:13:30 cloud_vm_ray_backend.py:888] To view detailed progress: tail -n100 -f /Users/zongheng/sky_logs/sky-2022-08-17-13-13-30-385888/provision.log
```
as well as reading
```
I 08-17 13:13:32 cloud_vm_ray_backend.py:1092] Launching on GCP us-central1 (us-central1-a)
I 08-17 13:14:26 log_utils.py:45] Head node is up.
I 08-17 13:15:58 cloud_vm_ray_backend.py:980] Successfully provisioned or found existing VM.
```

**Breakdown, with this PR**

- `_add_auth_to_cluster_config()`: ~1.5-2s
- **Launch VM till SSH ready: ~55s-1min**
  - Includes: check head node first + provision & wait for ssh + update status
  - [ ] Note that ray autoscale's "update status" using tags is expensive, 4-7s
- **Preparing SkyPilot runtime: ~1min 30s**
  - **File mounts with 3 clouds credentials: ~30s**
    - see the timestamp between `[2/7] Processing file mounts` to `[3/7] No worker file mounts to sync`
  - **Setup_commands: ~32s**
    - between `[6/7] Running setup commands` and `[7/7] Starting the Ray runtime`
   - **Starting Ray: ~19s**
     - between `[7/7] Starting the Ray runtime` and last line of provision.log
 - some other overheads (e.g., update-status 6 times!)
   
# Breakdown of optimizations
- Optimize interactive nodes: saves 6s (no autostop=-1), 9s (no run='').
- in cloud_vm_ray_backend 
  - (Saves 4s) Do not query IP for 1-node clusters; get head IP from `ray up` stdout
  - (Saves 4s) Do not query zone for 1-node, GCP clusters
  - (Saves 9s) Do not _ensure_cluster_ray_started() for fresh clusters
- in gcp-ray.yml.j2 
  - (Saves 10s) --disable-usage-stats in `ray start`
  - (Saves 7s) shorten start_commands from 3 to 2 SSH connections
  - (not much change) shorten setup_commands from 3 to 1 SSH connection
- [ ] (NOT INCLUDED in this PR) skip credentials in file_mounts: 30s

# What to optimize in the future 
1. file_mounts, see above.
2. Port these to other clouds.


# Tested
- [ ] smoke tests, passed except unrelated failures
  - `TestStorageWithCredentials`: failed on master for me
  - `test_spot_storage`: known access denied
  - `test_spot`, `test_use_spot`: AWS related, tracked in #1026 -- they passed when using a non-problematic region