@Michaelvll 2022-08-18
Maybe another PR: in the spot case, we have to run the `ray up` again to set the `upscaling_speed` to 0 to avoid ray spin up another cluster after the instance is preempted. Probably, we can exclude the single node case here, as in that case, the preempted node should always be the head node and will not be able to spin up another node.
https://github.com/skypilot-org/skypilot/blob/619655258e7b1479e5144d5a364257ac5dba5e4d/sky/backends/cloud_vm_ray_backend.py#L1173-L1182